requirements.txt
# Core Framework
streamlit>=1.29.0

# AI/LLM
groq>=0.4.1

# Database - Using psycopg (version 3) for better Python 3.13 compatibility
psycopg[binary]>=3.1.0
SQLAlchemy>=2.0.23

# YouTube API
google-api-python-client>=2.108.0

# Environment Management
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
pydantic>=2.5.0
